Implement reusable `FetchAgentKey`
